refactor: 💡 컨벤션 준수를 위한 패키지 이동 및 어노테이션 수정
- OAuthUserResponse 클래스의 패키지 이동, 어노테이션을 컨벤션에 맞게 수정
- 인증과 관련된 요청과 응답 DTO가 포함된 패키지를 controller 하위로 이동

✅ Closes: #40
