+++ src/main/java/com/postsquad/scoup/web/auth/GitHubOAuth.java
@@ -3,9 +3,9 @@ package com.postsquad.scoup.web.auth;
 import com.postsquad.scoup.web.auth.exception.AccessTokenNotFoundException;
 import com.postsquad.scoup.web.auth.exception.GitHubUserNotFoundException;
 import com.postsquad.scoup.web.auth.exception.GitHubRequestNotValidException;
-import com.postsquad.scoup.web.auth.request.AccessTokenRequest;
-import com.postsquad.scoup.web.auth.response.AccessTokenResponse;
-import com.postsquad.scoup.web.auth.response.OAuthUserResponse;
+import com.postsquad.scoup.web.auth.controller.request.AccessTokenRequest;
+import com.postsquad.scoup.web.auth.controller.response.AccessTokenResponse;
+import com.postsquad.scoup.web.auth.controller.response.OAuthUserResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
