+++ src/main/java/com/postsquad/scoup/web/auth/controller/OAuthController.java
@@ -1,6 +1,6 @@
 package com.postsquad.scoup.web.auth.controller;
 
-import com.postsquad.scoup.web.user.controller.response.SocialAuthenticationResponse;
+import com.postsquad.scoup.web.auth.controller.response.SocialAuthenticationResponse;
 import com.postsquad.scoup.web.auth.service.OAuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
