+++ src/main/java/com/postsquad/scoup/web/auth/controller/response/SocialAuthenticationResponse.java
@@ -1,12 +1,8 @@
-package com.postsquad.scoup.web.user.controller.response;
+package com.postsquad.scoup.web.auth.controller.response;
 
-import com.postsquad.scoup.web.auth.response.OAuthUserResponse;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
-@Getter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
